SecretHitler - states and rejoin
Started implementing the player rejoin functionality, so if a player reconnects then the socket will emit to that player the current game state, so they can rejoin at the correct point in the game. Also added message when the "start game" button is clicked if too few users are connected.