Create bot-heroku.js
Created bot-heroku.js 
This code has been updated to play nice with Heroku (just about). For some reason it needs the "http" module so that I can specify the port for Heroku, and a little code added to log the number of games played for my reference. This is the version now live and available publicly.
Minor tweak - guesses no longer need to be capitalised.