Working towards building game functionality
New cluesInput component for entering guesses, roundsSection displays clues as they come in (ugly as sin but working). Want to change clue functionality (client/server) to show the name of the person submitting the guess. Also a button to switch from submitting a clue to a solution. Then onto thumbs/win functionality.